When writing JSON, write integers and floats as such (not strings)
When writing a JSON file for example when exporting database
content into JSON format, write integers and floats as numeric
values into the JSON file instead of as strings.

String values of course work as well, but since JSON supports the
correct data types for these values, would be better to also
export them as such.