Update adminerTablesFilter to set up filter asap
If you're opening a very large database on a very slow server,
adminer might take a few seconds to load the whole table list.

Because the JS for the filter is added after the tables are
shown this can result in the user typing into the filter and
nothing happening, followed by the filter contents being replaced
by the sessionStorage cached contents.

This is confusing and annoying, so this commit does most of the
work beforehand and only leaves triggering the filtering itself
until after the whole list of tables has been printed.