Make the ME constant pluggable
When creating an integration of Adminer as a third party service, in a
similar manner to PHPMyAdmin, query string parameters used for
selection/authentication are destroyed. This hinders the ability to
integrate Adminer when robust, non-rewritten URLs are used.

This constant enables preservation of parameters via a simple define:

define('MEQPFX', 'some_token_param=' . rawurlencode($some_token_value) . '&');