Elastic: Properly display sparse result rows
Result records in Elasticsearch do not always have all columns
that are defined in an index.
This often happens when multiple document types are stored in the same index.

The first row has columns ["_id", "html", "url"], while the second
misses the "html" column: ["_id", "url"].

Up to now, Adminer expected that all result rows include all columns.
This led to the problem that the "url" value in the 2nd example row
was rendered in the "html" column.

This patch fixes this problem by iterating over the expected keys,
and filling NULL when the database result row misses this key.

(I tried to fix that in the Elasticsearch driver only, but this
is much more complicated, since MinDriver::select() does not know
about expected field names at all.)